03/22/2021
Messy. Hectic. But.....consistent? We’ll see what day 20 nets but day 12 and day 15 should also be pretty huge tbh.

If we’re playing the continuous game, technically it should be only 5/100 that make the most impact. Could today be one of those days? God willing. Let’s do this....but....just show up tomorrow and that’s enough too.