Found a cool app store marketing / education value-hack
Having a profile at the end does make it feel a lot more personable for me. Usually, it's just some CTA or feature desc. Figured it would help devs looking to score a win with a specific crowd! Meaning: Showing who's behind it might be useful for somethings and not for others. Be mindful of fatigue and also down funnel use of app images in ads (& apple policy). But yeah, idk tho...i don't have an app so I haven't "made this decision" haha