Do not use for destruction or menace. Please. Thank you.
Do not expect me to actually follow up or build on top of this. I need to first see how much risk, danger, and division sharing this already brings to the world. I’m an engineer and more than God where I’m at. Please be mindful and kind in seeing my honest and creative intentions. If you think I’ll risk the other 37 days for an idea that’s not my main focus... you will be disappointed. I’m not a message, symbol, your hero, or even a means to XYZ. I’m just showing up for all of us for 100. I’m always scared too. If I’m not, it’s likely a signal to check myself. Fast. Thank you and I appreciate you